+++ views/users/index.pug
@@ -1,4 +1,6 @@
 h1 Users
+a(href='/') Home
+br
 a(href='/users/create') New
 br
 br
